feat(audio): use Float64 throughout the DSP pipeline
Promote audio transport, conversion, resampling, mixing, equalization,
fades, HRTF, FreeSurround, Signalsmith, and output processing to
double precision.

Preserve native PCM and DoP passthrough where possible, improve buffer
and format-transition safety, and convert to Float32 only at external
APIs or device boundaries that require it.

+++ Audio/Chain/DSP/DSPFaderNode.h
@@ -34,7 +34,7 @@
 - (void)waitForReset;
 - (void)setDoPMode:(BOOL)enabled;
 
-- (float)fadeLevel;
+- (double)fadeLevel;
 
 - (void)appendFadeOut:(FadedBuffer *_Nonnull)buffer;
